# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorField"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorField": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ErrorField",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionRefreshResponse": {
        "examples": [
          {
            "access_token_expires_at": "2026-08-11T09:15:00Z",
            "authenticated_at": "2026-08-11T09:00:00Z",
            "refresh_token_expires_at": "2026-08-25T09:00:00Z"
          }
        ],
        "properties": {
          "access_token_expires_at": {
            "format": "date-time",
            "title": "Access Token Expires At",
            "type": "string"
          },
          "authenticated_at": {
            "format": "date-time",
            "title": "Authenticated At",
            "type": "string"
          },
          "refresh_token_expires_at": {
            "format": "date-time",
            "title": "Refresh Token Expires At",
            "type": "string"
          }
        },
        "required": [
          "authenticated_at",
          "access_token_expires_at",
          "refresh_token_expires_at"
        ],
        "title": "SessionRefreshResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessCookieAuth": {
        "description": "HttpOnly access JWT 쿠키",
        "in": "cookie",
        "name": "yeongyangkkuk_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "영양꾹 1차 로컬 MVP 백엔드 API",
    "title": "영양꾹 백엔드",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "yeongyangkkuk_refresh_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yeongyangkkuk Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "현재 세션 폐기 또는 이미 로그아웃 상태",
            "headers": {
              "Cache-Control": {
                "description": "인증 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "회전 또는 삭제되는 access·refresh HttpOnly 쿠키",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 처리 실패"
          }
        },
        "summary": "현재 기기 로그아웃",
        "tags": [
          "인증"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "auth_refresh_session",
        "parameters": [
          {
            "in": "cookie",
            "name": "yeongyangkkuk_refresh_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yeongyangkkuk Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefreshResponse"
                }
              }
            },
            "description": "현재 refresh 세션 회전 성공",
            "headers": {
              "Cache-Control": {
                "description": "인증 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "회전 또는 삭제되는 access·refresh HttpOnly 쿠키",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_SESSION_INVALID",
                    "fields": [],
                    "message": "유효하지 않은 인증 세션입니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "refresh 세션 무효",
            "headers": {
              "Cache-Control": {
                "description": "인증 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "회전 또는 삭제되는 access·refresh HttpOnly 쿠키",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 처리 실패"
          }
        },
        "summary": "인증 세션 갱신",
        "tags": [
          "인증"
        ]
      }
    },
...
  }
}